\name{print.ggcolpairs}
\alias{print.ggcolpairs}
\title{Print ggcolpairs}
\author{Fernando Fuentes \email{fuentes.f4@gmail.com}}
\keyword{internal}
\description{
Print function for ggcolpairs. Plots matrix on current graphics device.
}
\usage{print.ggcolpairs(colpairs, ...)}
\arguments{
\item{colpairs}{ggcolpairs object to display}
\item{...}{not used}
}
